+++ packages/react-sdk/src/core/components/StreamVideo/StreamVideo.tsx
@@ -1,0 +1,12 @@
+import {
+  StreamVideoProvider,
+  StreamVideoProps,
+} from '@stream-io/video-react-bindings';
+import { PropsWithChildren } from 'react';
+import { translations } from '../../../translations';
+
+export const StreamVideo = (props: PropsWithChildren<StreamVideoProps>) => {
+  return (
+    <StreamVideoProvider translationsOverrides={translations} {...props} />
+  );
+};